UI Scaling + GameOver/WinGame Sounds
Instead of using the soundmanager for game win and over, the scenes themselves run a quick oneshot of the sound it needs to play in the audio source via the editor.

All the UI when built is scaled incorrectly, so this has been fixed through the editor by adjusting each canvas to >scale with screen, and then editing the values of the screen size to 1920x1080.